+++ modules/castbar/init.lua
@@ -1,6 +1,6 @@
 --[[
     VUI - Castbar Module
-    Version: 0.0.1
+    Version: 1.0.0
     Author: VortexQ8
 ]]